feat: resume interrupted network/WSL drive scans instead of restarting from scratch
Network and WSL drive indexing previously always walked the whole tree from the
root on every refresh, so an interrupted scan (or even a routine periodic one)
paid the full network cost again every time. TreeBuilder now diffs against the
previous index: a directory whose own LastWriteTime still matches what was
recorded, and that was fully enumerated last time (FileRecordFlags.Listed),
has its cached children reused instead of being re-listed. A new
FileRecordStore.IsComplete flag distinguishes a genuinely finished index from
a mid-walk checkpoint, so an interrupted scan is recognized as unfinished and
resumed on the next refresh instead of being silently treated as done.

Along the way, fixed three bugs this surfaced or that blocked it entirely:
- A race where a directory's own record wasn't registered before another
  worker could dequeue and finish it, silently leaving it unmarked forever.
- RuntimeIndex stored per-record flags in a byte, one bit short of the new
  Listed flag -- every directory's mark was silently truncated away on load.
- A transient file lock during a concurrent checkpoint save was misread as a
  corrupted cache and deleted outright instead of just being retried later.

Also: the root directory's own mtime is now tracked (was hardcoded to 0,
permanently forcing a share's top-level entries to be re-listed every time),
checkpointing is strictly count-based instead of also firing on a wall-clock
timer, and diagnostic logging reports resume progress and total scan time.

+++ Core/SearchIndex/RecordIndex/HardLinkUpdateExtensions.cs
@@ -95,7 +95,7 @@ internal static class HardLinkUpdateExtensions
             return;
 
         var wasDirectory = index.IsDirectory(idx);
-        index.Flags[idx] = (byte)(((FileRecordFlags)index.Flags[idx]) | FileRecordFlags.Deleted);
+        index.Flags[idx] = (ushort)(((FileRecordFlags)index.Flags[idx]) | FileRecordFlags.Deleted);
         index.CharMasks[idx] = 0;
 
         var parentIndex = index.ParentIndexes[idx];
